+++ code/crates/core-consensus/src/handle/proposed_value.rs
@@ -81,6 +81,7 @@ where
         perform!(
             co,
             Effect::WalAppend(
+                proposed_value.height,
                 WalEntry::ProposedValue(proposed_value.clone()),
                 Default::default()
             )